+++ tests/e2e/expected_download.py
@@ -119,7 +119,7 @@ def assert_expected_downloads(
     dry_run: bool,
     expected_download_summary_file_name: str,
     ignore_md5_hashes_for: Optional[List[str]] = None,
-    regenerate_expected_download_summary: bool = True,
+    regenerate_expected_download_summary: bool = False,
 ):
     if dry_run:
         output_directory_contents = list(Path(output_directory).rglob("*"))
