+++ tests/e2e/expected_transaction_log.py
@@ -10,7 +10,7 @@ def assert_transaction_log_matches(
     output_directory: str,
     transaction_log: FileHandlerTransactionLog,
     transaction_log_summary_file_name: str,
-    regenerate_transaction_log: bool = True,
+    regenerate_transaction_log: bool = False,
 ):
     """
     Parameters
